Assignment3/api/requirements.txt
uvicorn>=0.18.2, <0.19.0
fastapi>=0.79.0, <1.0.0
python-multipart>=0.0.5, <0.1.0
pydantic>=1.9.1, <1.10.0
typing_extensions>=3.10.0, <3.11.0
loguru>=0.6.0, <0.7.0
joblib==1.2.0
numpy==1.19.5
pandas==1.2.5
scipy==1.6.2
matplotlib==3.6.2
scikit_learn==0.22.1
xarray==0.20.1
Cython==0.29.32

python-jose==3.3.0
passlib==1.7.4
gunicorn==20.1.0